local/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
flask>=3.0.0
cryptography>=41.0.0
pycryptodome>=3.19.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
openai>=1.0.0

